\name{OUwie-class}
\docType{class}
\title{OUwie object}
\alias{OUwie-class}
\description{
  A fitted generalized Hansen model object.
}
\details{
  The function \code{OUwie} creates a \code{OUwie} object by fitting a generalized Hansen model.
}
\author{Jeremy M. Beaulieu and Brian C. O'Meara}
\keyword{models}